Manage cache file race conditions
Checking cache_file existence and opening it is not an atomic operation
and can lead to errors if cache file has been removed in the meantime.

In the same way, we can't be sure that the cache file exists when we
delete it.

Handle exceptions on opening and loading JSON file:
- For IO errors (file probably does not exist), return None
- For any other error (invalid JSON or else), or when cache is expired,
  remove the cache file and return None

Signed-off-by: Nicolas Payart <npayart@gmail.com>