feat: add specific code path for AWX
The ansible bastion wrapper default lookup to get bastion vars, is to run "ansible-inventory --list"
A downside of this way of doing, is that the whole inventory is going to be evaluated, and if we are using some custom vars plugins, there are executed to.
It can end up being very time consuming.

When using AWX, the whole inventory is available in the AWX Execution Environment as a file.
So it is much easier and faster to get the host associated to the requested ip (the ip sent by Ansible to the ssh wrapper).
Then, we can look for the bastion vars in the host vars, and if not found, execute "ansible-inventory --host" on the specific host, instead of the whole inventory.